requirements.txt
ell==0.16.0
bs4==0.0.1
lxml==4.9.1
requests==2.27.1
python-socketio==5.8.0